Add SceneObject class and detect_collision property on player
SceneObjects have an ID, a 'solid' flag, a name/tag, x and y
coordinates, width and heights, a sprite and visibility

The detect_collision method checks for each object in an object list,
and checks if it's solid and the hitbox of the player collides with the
object's hitbox. If so, returns the object.